requirements.txt
playwright
groq
rich
python-dotenv
